requirements.txt
uvicorn[standard]
gunicorn
fastapi
google-cloud-secret-manager
openai
python-multipart
